+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
 - Added `add-docs`, `update-docs` commands to `pyrefdev-indexer`.
+- Added missing module symbols.
 - Support [boto3](https://boto3.amazonaws.com/v1/documentation/api/latest/index.html).
 - Support [setuptools](https://setuptools.pypa.io/en/latest/).
 - Support [botocore](https://botocore.amazonaws.com/v1/documentation/api/latest/index.html).